Move deterministic QC into the pipeline
structure.md puts room_format at "what we can do with a Room.py folder" — export
Blender/GLB, render videos, export the viewer — and puts QC in the pipeline: "Last
comes deterministic QC over the resulting Room.py ... So all the QC lives in this
folder too." room_format/validation was neither, and nothing else in room_format
imported it, so it lifts out cleanly.

The report and the resolvers are the QC, and move to pipeline/qc. The box arithmetic
underneath them is not a report — it is what check_collisions runs on every agent
turn — so it moves under the tool that owns it, per the source-under-its-tool
convention. pipeline/qc imports it back, which is the direction structure.md
intends ("we can also reuse them outside the agentic loop by calling into
agent/tools"); the reverse would have made agent depend on pipeline, which it
does not do anywhere today.

  pipeline/qc/checks.py                     <- validation/room.py (qc + CLI)
  pipeline/qc/{fix,correct}.py              <- validation/{fix,correct}.py
  .../check_collisions/source/geometry.py   <- validation/room.py (primitives)
  .../check_collisions/source/collision_mesh.py

Drops the unused _overlap helper.
