+++ src/litereality_agent/agent/tools/check_collisions/source/__init__.py
@@ -1,0 +1,1 @@
+"""Source the `check_collisions` tool runs on."""
