+++ src/litereality_agent/agent/tools/check_collisions/source/collision_mesh.py
