+++ src/litereality_agent/pipeline/qc/correct.py
@@ -16,8 +16,8 @@ are anchored to walls (needs a human / the model pass).
 
 Needs a compiled `Room.glb` (build the room first) and `python-fcl`.
 
-    python -m litereality_agent.room_format.validation.correct --room <dir>          # dry run: print the plan
-    python -m litereality_agent.room_format.validation.correct --room <dir> --apply  # write the nudges into Room.py
+    python -m litereality_agent.pipeline.qc.correct --room <dir>          # dry run: print the plan
+    python -m litereality_agent.pipeline.qc.correct --room <dir> --apply  # write the nudges into Room.py
 """
 
 from __future__ import annotations
@@ -27,8 +27,9 @@ import glob
 import math
 from pathlib import Path
 
-from litereality_agent.room_format.validation import collision_mesh as sc
-from litereality_agent.room_format.validation.fix import (
+from litereality_agent.agent.tools.check_collisions.source import collision_mesh as sc
+from litereality_agent.agent.tools.check_collisions.source.geometry import _extract_shell
+from litereality_agent.pipeline.qc.fix import (
     FREE_STANDING,
     MAX_NUDGE,
     _anchored,
@@ -36,7 +37,6 @@ from litereality_agent.room_format.validation.fix import (
     _replace_center,
     _validate,
 )
-from litereality_agent.room_format.validation.room import _extract_shell
 
 # scene_collision findings that come with a metric fix we can apply deterministically.
 _FIXABLE = ("object_clash", "wall_clash")
